Fix Updating ControlNet Properties
Issue:
In the lamba assinged to the slot 'widget.qcombo.editTextChanged.connect'
We take a copy of the default config into the lambda here:
'lambda value: self.set_input(>>inputs<<, key, value)'
This mean we are only able to change one property at a time as the other are reset to the default.

Fix:
We need to get the current config inside the lamba before we modify and update it.
not sure if this is the most optimal fix but it does fix the issue.

dict(script.cfg(f"controlnet{self.unit}_inputs", dict))

Bad Case:
update: detect_hand to:disable
{'coarse': 'disable', 'detect_body': 'enable', 'detect_face': 'enable', 'detect_hand': 'disable', 'max_faces': 10, 'min_confidence': 0.5, 'safe': 'enable', 'threshold': 0, 'version': 'v1'}
update: detect_face to:disable
{'coarse': 'disable', 'detect_body': 'enable', 'detect_face': 'disable', 'detect_hand': 'enable', 'max_faces': 10, 'min_confidence': 0.5, 'safe': 'enable', 'threshold': 0, 'version': 'v1'}