Prepend paths to known Lora
Issue:
SD automatic uses the name only lora format
<lora:people_count_slider_v1:0.5>
Comfy uses the full path
<lora:OstrisUtils/people_count_slider_v1:0.5>

Fix:
We collect all the know loras from the server on startup
Then we attempt to match the lora name to the known lora paths when creating the workflow.
This allows for use of the same lora format as SD automatic
if there are multiple loras that could match we use the original string.

- Updated Lora regex to allow for negative weights.